% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_csv.R
\name{tar_csv}
\alias{tar_csv}
\title{Save a data frame as a csv}
\usage{
tar_csv(dataframe_target, export_path = "exports")
}
\arguments{
\item{dataframe_target}{the name of the target that will generate the data frame for export}

\item{export_path}{the destination path of the exported csv file}
}
\value{
A target object for exporting a plot as a png.
}
\description{
A target for exporting a data frame as a csv file.
}
\details{
The name of the returned target will be the name of the data frame target with the prefix \code{export_}.
The file name of the exported csv file will be that of the argument provided to \code{dataframe_target}.
}
\examples{
if (identical(Sys.getenv("TAR_LONG_EXAMPLES"), "true")) {
  targets::tar_dir({
    targets::tar_script({
      list(
        targets::tar_target(
          iris,
          iris
        ),
        jftargets::tar_csv(iris)
      )
    })
    
    targets::tar_make()
    tar_read(export_iris)
  })
}
}
