% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_png.R
\name{tar_png}
\alias{tar_png}
\title{Save a plot as a png}
\usage{
tar_png(
  plot_target,
  width = 9,
  height = 5,
  dpi = 300,
  units = c("in", "cm", "mm", "px"),
  export_path = "exports/figures"
)
}
\arguments{
\item{plot_target}{the name of the target that will generate the plot to export}

\item{width, height, units}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{export_path}{the destination path of the exported plot file}
}
\value{
A target object for exporting a plot as a png.
}
\description{
A target for exporting a plot target as a png.
}
\details{
The name of the returned target will be the name of the plot target with the prefix \code{export_}.
The file name of the exported plot will be that of the argment provided to \code{plot_target}.
}
\examples{
if (identical(Sys.getenv("TAR_LONG_EXAMPLES"), "true")) {
  targets::tar_dir({
    targets::tar_script({
      list(
        targets::tar_target(
          iris_plot,
          ggplot2::ggplot(
            iris,
            ggplot2::aes(
              x = Sepal.Length,
              y = Sepal.Width
            ) 
          ) +
            ggplot2::geom_point()
        ),
        jftargets::tar_png(iris_plot)
      )
    })
    
    targets::tar_make()
    tar_read(export_iris_plot)
  })
}
}
