% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_xlsx.R
\name{tar_xlsx}
\alias{tar_xlsx}
\title{Save a data frame as an xlsx}
\usage{
tar_xlsx(dataframe_target, export_path = "exports")
}
\arguments{
\item{dataframe_target}{the name of the target that will generate the data frame for export}

\item{export_path}{the destination path of the exported xlsx file}
}
\value{
A target object for exporting a plot as a png.
}
\description{
A target for exporting a data frame as a xlsx file.
}
\details{
The name of the returned target will be the name of the data frame target with the prefix \code{export_}.
The file name of the exported xlsx file will be that of the argument provided to \code{dataframe_target}.
}
\examples{
if (identical(Sys.getenv("TAR_LONG_EXAMPLES"), "true")) {
  targets::tar_dir({
    targets::tar_script({
      list(
        targets::tar_target(
          iris,
          iris
        ),
        jftargets::tar_xlsx(iris)
      )
    })
    
    targets::tar_make()
    tar_read(export_iris)
  })
}
}
